Broke out cpu/cpu.go into different files
I've moved all the register functions into `register.go`, the stack
functions into `stack.go` and the utilities into `util.go`.

There are now tests for the registers / stack, but none for the CPU.
To get 100% coverage of the CPU would require some time.  Perhaps
in the future.

For the moment I'm happy that this is cleaner and I'm going to
close #4.