Improved error-handling.
It is a gross bug for the interpreter section of our code to call
os.Exit().   To resolve that I've updated the CPU object to return
an error from `Run`, and the registers will return errors if their
contents have the wrong types.

I'll update the fuzzer for Go 1.18+ to add confidence here too.